auto-add connection-start time offsets
if the number of clients is high relative to the request rate,
mcshredder tends to run all of the traffic at 100% speed, then sleep for
the rest of the period.

this ensures that the first request from each client connection is
spaced out across the specified period. Each sub request is delta'd as
per the rate limit.