mcs.match(req, res) to check value matches key
Values are generated from the hash of their original key. Responses are then
checked against the hash of the key of the original request.

This should generally find if the response matches the request, but it
cannot tell if values update properly for the same key. Will need
something different for that, or some way to vary the value hash.

tries to be quick about it.