experimental request factory code
also adds the start of an alloc profiler.

The goal is to pair { init = true } with request factories to reduce the
amount of runtime allocation and improve performance and testing
stability.

Coming soon will be cached response objects and inverting string
comparisons to avoid string copies _into_ lua. With these changes it
should be possible to have zero-alloc runtimes for most tests.

Need to fill in a lot of factory request code still; will likely use
unions to type the req_factory struct and allow things like varying
parameters.